+++ Filminurk/Models/Movies/MoviesIndexViewModel.cs
@@ -6,7 +6,7 @@
         public string Title { get; set; }
         
         public DateOnly FirstPublished { get; set; }
-        public decimal? CurrentRating { get; set; }
+        public double? CurrentRating { get; set; }
         //public List<UserComment>? Reviews { get; set; }
         public int? OscarsWon { get; set; }
         public decimal? IMDbRating { get; set; }
